. Implement callback after TalkTo . Guard advise player x seconds after topic window closed, not after start talkTo . Fix bug about DisplayCollisionEntities() . Fix bug about collisionMap was generated in the tmp directory after map saving . Fix bug about previous scene was not cleared when starting to edit a new scene . Fix multiple collision bugs, like character sinks into the ground and go under some walls . Fix runtime issue in VM about jne/jae/jbe because the jmp adress was get into 1 byte instead of 4

+++ EasyEngineV4/EasyEngineSrc/Collision/CollisionManager.cpp
@@ -197,7 +197,7 @@ void CCollisionManager::CreateHeightMapWithoutRender(string sFileName)
 	m_oLoaderManager.Load(sFileName, ami);
 	ILoader::CMeshInfos& mi = ami.m_vMeshes[0];
 	CVector dim = mi.m_pBoundingBox->GetDimension();
-	int nCellCount = dim.m_x * dim.m_z;
+	int nCellCount = (int)(dim.m_x * dim.m_z);
 	int quadCount = mi.m_vIndex.size() / 6;
 	set<float> setx;
 	set<float> sety;
